Refactor server implementation
